Med: ais: Get the CRM to the point where it mostly functions under openais A single node can start and become the DC Implemented remote client status and integrated it into the membership updates Automated the processing of membership updates into crm_peer_cache
Still need to impliement quorum but all the pieces are there to do so

Mercurial revision: e8dda2212f2c951d5f23a38977b8fe12be1b034d

+++ crm/crmd/election.c
@@ -293,8 +293,8 @@ do_election_count_vote(long long action,
 	} else {
 		CRM_CHECK(strcasecmp(fsa_our_uname, vote_from) != 0, ;);
 		crm_debug("Them: %s (born=%llu)  Us: %s (born=%llu)",
-			  vote_from, your_node->born,
-			  fsa_our_uname, our_node->born);
+			  vote_from, (unsigned long long)your_node->born,
+			  fsa_our_uname, (unsigned long long)our_node->born);
 /* cant happen...
  *	} else if(strcasecmp(fsa_our_uname, vote_from) == 0) {
  *
